Split homepage into multiple components
