+++ app/src/components/interface/footer/Footer.component.jsx
@@ -1,0 +1,148 @@
+import {
+    Button,
+    Center,
+    Container,
+    HStack,
+    Image,
+    Link,
+    Text,
+    useColorMode,
+    VStack
+} from '@chakra-ui/react';
+
+import logodark from 'images/logodark.png';
+import logolight from 'images/logolight.png';
+import { observer } from 'mobx-react';
+import { useContext } from 'react';
+import { RootStoreContext } from 'stores/RootStore';
+
+function Footer() {
+    const { colorMode } = useColorMode();
+    const store = useContext(RootStoreContext);
+
+    return (
+        <Container
+            maxW="container.xl"
+            justifyContent="space-evenly"
+            display="flex"
+        >
+            <Center
+                paddingTop="100px"
+                paddingBottom="50px"
+                maxWidth="300px"
+                flexDir="column"
+                alignItems="start"
+            >
+                <HStack justifyContent="center" marginBottom="20px">
+                    <Image
+                        src={logodark}
+                        alt="Collaboration spotting logo"
+                        height="20px"
+                        display={colorMode === 'light' ? 'none' : 'block'}
+                    />
+                    <Image
+                        src={logolight}
+                        alt="Collaboration spotting logo"
+                        height="20px"
+                        display={colorMode === 'light' ? 'block' : 'none'}
+                    />{' '}
+                    <Text fontWeight="bold">Collaboration Spotting X</Text>
+                </HStack>
+                <Text marginBottom="20px" textAlign="left" fontSize="xs">
+                    Developed at <b>CERN</b>, Geneva, Switzerland by{' '}
+                    <b>Aleksandar Bobić</b> led by <b>Dr. Jean-Marie Le Goff</b>{' '}
+                    and <b>prof. Christian Gütl</b>.
+                </Text>
+
+                <Text
+                    fontStyle="italic"
+                    fontSize="xs"
+                    textAlign="left"
+                    marginBottom="20px"
+                >
+                    This project was inspired by the{' '}
+                    <Link
+                        fontWeight="bold"
+                        textDecoration="underline"
+                        display="inline"
+                        opacity="0.75"
+                        target="_blank"
+                        href="https://collaborationspotting.web.cern.ch/"
+                        _hover={{ opacity: 1 }}
+                    >
+                        Collaboration Spotting project
+                    </Link>
+                    . We would like to thank the{' '}
+                    <Link
+                        fontWeight="bold"
+                        textDecoration="underline"
+                        display="inline"
+                        opacity="0.75"
+                        target="_blank"
+                        href="https://ercim-news.ercim.eu/en111/r-i/collaboration-spotting-a-visual-analytics-platform-to-assist-knowledge-discovery"
+                        _hover={{ opacity: 1 }}
+                    >
+                        Collaboration Spotting team
+                    </Link>{' '}
+                    for their contributions.
+                </Text>
+            </Center>
+            <Center maxWidth="300px">
+                <VStack alignItems="start">
+                    <Link
+                        fontWeight="bold"
+                        fontSize="sm"
+                        textDecoration="underline"
+                        display="inline"
+                        opacity="0.75"
+                        target="_blank"
+                        href="https://github.com/aleksbobic/csx"
+                        _hover={{ opacity: 1 }}
+                    >
+                        Github
+                    </Link>
+                    <Link
+                        fontWeight="bold"
+                        textDecoration="underline"
+                        fontSize="sm"
+                        display="inline"
+                        opacity="0.75"
+                        target="_blank"
+                        href="https://csxp.me"
+                        _hover={{ opacity: 1 }}
+                    >
+                        Webpage
+                    </Link>
+                    <Button
+                        variant="unstyled"
+                        size="sm"
+                        padding="0"
+                        margin="0"
+                        fontSize="sm"
+                        textDecoration="underline"
+                        opacity="0.75"
+                        fontWeight="bold"
+                        _hover={{ opacity: 1 }}
+                        height="21px"
+                        onClick={() => {
+                            store.search.setSearchIsEmpty(false);
+                            store.core.setShowCookieInfo(true);
+                        }}
+                    >
+                        Cookies & local storage
+                    </Button>
+                    <Text
+                        textAlign="left"
+                        fontWeight="bold"
+                        fontSize="xs"
+                        paddingTop="16px"
+                    >
+                        CERN &copy; 2022
+                    </Text>
+                </VStack>
+            </Center>
+        </Container>
+    );
+}
+
+export default observer(Footer);
